Carrera distancia done

+++ app/src/main/java/com/cuatrodivinas/seekandsolve/Datos/CarreraActual.kt
@@ -5,9 +5,9 @@ import java.io.Serializable
 class CarreraActual(
     val idCarrera: String,
     var acertijosPrimerIntento: Int,
-    var distanciaRecorrida: Int,
+    var distanciaRecorrida: Double,
     var puntosCompletados: MutableList<Punto>
 ): Serializable {
 
-    constructor(): this("", 0, 0, mutableListOf())
+    constructor(): this("", 0, 0.0, mutableListOf())
 }
